Seperetaed the Async and Sync tests
