+++ Zoo/Views/Shared/_NavBar.cshtml
@@ -36,8 +36,8 @@
                         </a>
                     </li>
                     <li class="nav-item active me-1">
-                        <a asp-action="SignUp" asp-controller="User" class="nav-link active fw-bolder">
-                            Sign Up
+                        <a asp-action="Register" asp-controller="User" class="nav-link active fw-bolder">
+                            Register
                         </a>
                     </li>
                 }
